# src/main/resources/swagger/scad.yaml
openapi: 3.0.3
info:
  title: Biblioteca Digital API
  description: API para o cadastro de autores e livros na Biblioteca Digital.
  version: 1.0.0

tags:
  - name: Autores
    description: Operações relacionadas a autores
  - name: Livros
    description: Operações relacionadas a livros

paths:
  /autores:
    post:
      tags:
        - Autores
      summary: Cadastra um novo autor
      operationId: registerAuthor
      requestBody:
        description: Objeto com os dados do autor a ser criado.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AutorRequest'
        required: true
      responses:
        '201':
          description: Autor cadastrado com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AutorResponse'
        '400':
          description: Requisição inválida.
        '422':
          description: Erro de validação.
        default:
          description: Erro inesperado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      tags:
        - Autores
      summary: Lista autores com filtros opcionais
      operationId: getAllAuthors
      parameters:
        - name: nacionalityCountry
          in: query
          description: Filtra autores pelo país de nacionalidade.
          required: false
          schema:
            type: string
            example: Brasil
        - name: name
          in: query
          description: Filtra autores cujo nome contém o valor fornecido.
          required: false
          schema:
            type: string
            example: Maria
      responses:
        '200':
          description: Lista de autores retornada com sucesso.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AutorResponse'
        default:
          description: Erro inesperado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /autores/{id}:
    parameters:
      - name: id
        in: path
        description: ID do autor.
        required: true
        schema:
          type: integer
          format: int64
    get:
      tags:
        - Autores
      summary: Retorna um autor específico
      operationId: getAuthorById
      responses:
        '200':
          description: Autor encontrado com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AutorResponse'
        '404':
          description: Autor não encontrado.
        default:
          description: Erro inesperado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      tags:
        - Autores
      summary: Atualiza os dados de um autor
      operationId: updateAuthor
      requestBody:
        description: Objeto com os dados do autor a ser atualizado.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AutorRequest'
      responses:
        '200':
          description: Autor atualizado com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AutorResponse'
        '404':
          description: Autor não encontrado.
        '422':
          description: Erro de validação.
        default:
          description: Erro inesperado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags:
        - Autores
      summary: Deleta um autor
      operationId: deleteAuthor
      responses:
        '204':
          description: Autor deletado com sucesso.
        '404':
          description: Autor não encontrado.
        default:
          description: Erro inesperado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /books:
    post:
      tags:
        - Livros
      summary: Cadastra um novo livro
      operationId: registerBook
      requestBody:
        description: Objeto com os dados do livro a ser criado.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookRequest'
        required: true
      responses:
        '201':
          description: Livro cadastrado com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookResponse'
        '400':
          description: Requisição inválida.
        '422':
          description: Erro de validação.
        default:
          description: Erro inesperado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      tags:
        - Livros
      summary: Lista livros com filtros opcionais
      operationId: getAllBooks
      parameters:
        - name: title
          in: query
          description: Filtra livros pelo título.
          required: false
          schema:
            type: string
            example: Dom Casmurro
        - name: isbn
          in: query
          description: Filtra livros pelo ISBN.
          required: false
          schema:
            type: string
            example: 978-85-359-0277-X
        - name: datePublisher
          in: query
          description: Filtra livros pela data de publicação.
          required: false
          schema:
            type: string
            format: date
            example: "1899-03-15"
      responses:
        '200':
          description: Lista de livros retornada com sucesso.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BookResponse'
        default:
          description: Erro inesperado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /books/{id}:
    parameters:
      - name: id
        in: path
        description: ID do livro.
        required: true
        schema:
          type: integer
          format: int64
    get:
      tags:
        - Livros
      summary: Retorna um livro específico
      operationId: getBookById
      responses:
        '200':
          description: Livro encontrado com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookResponse'
        '404':
          description: Livro não encontrado.
        default:
          description: Erro inesperado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      tags:
        - Livros
      summary: Atualiza os dados de um livro
      operationId: updateBook
      requestBody:
        description: Objeto com os dados do livro a ser atualizado.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookRequest'
      responses:
        '200':
          description: Livro atualizado com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookResponse'
        '404':
          description: Livro não encontrado.
        '422':
          description: Erro de validação.
        default:
          description: Erro inesperado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags:
        - Livros
      summary: Deleta um livro
      operationId: deleteBook
      responses:
        '204':
          description: Livro deletado com sucesso.
        '404':
          description: Livro não encontrado.
        default:
          description: Erro inesperado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    # === Schemas de Autor ===
    AutorRequest:
      type: object
      required:
        - name
        - dateBirth
        - nacionalityCountry
      properties:
        name:
          type: string
          description: Nome completo do autor.
          example: "Machado de Assis"
        dateBirth:
          type: string
          format: date
          description: Data de nascimento do autor.
          example: "1839-06-21"
        nacionalityCountry:
          type: string
          description: País de nacionalidade do autor.
          example: "Brasil"

    AutorResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID único do autor.
          example: 101
        name:
          type: string
          example: "Machado de Assis"
        dateBirth:
          type: string
          format: date
          example: "1839-06-21"
        nacionalityCountry:
          type: string
          example: "Brasil"
        books:
          type: array
          items:
            $ref: '#/components/schemas/BookResponse'

    # === Schemas de Livro ===
    BookRequest:
      type: object
      required:
        - title
        - isbn
        - datePublisher
        - authorId
        - price
        - genre
      properties:
        title:
          type: string
          description: Título do livro.
          example: "Dom Casmurro"
        isbn:
          type: string
          description: ISBN do livro.
          example: "978-85-359-0277-X"
        datePublisher:
          type: string
          format: date
          description: Data de publicação do livro.
          example: "1899-03-15"
        authorId:
          type: integer
          format: int64
          description: ID do autor do livro.
          example: 101
        price:
          type: number
          format: double
          description: Preço do livro.
          example: 45.90
        genre:
          $ref: '#/components/schemas/Genre'

    BookResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID único do livro.
          example: 201
        title:
          type: string
          example: "Dom Casmurro"
        isbn:
          type: string
          example: "978-85-359-0277-X"
        datePublisher:
          type: string
          format: date
          example: "1899-03-15"
        price:
          type: number
          format: double
          example: 45.90
        genre:
          $ref: '#/components/schemas/Genre'
        author:
          $ref: '#/components/schemas/AutorResponse'

    Genre:
      type: string
      description: Gênero literário do livro.
      enum:
        - ROMANCE
        - FICCAO_CIENTIFICA
        - FANTASIA
        - TERROR
        - AVENTURA
        - BIOGRAFIA
        - HISTORIA
        - ACTION
        - SUSPENSE

    # === Schema de Erro ===
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message